+++ src/main/java/com/holmes/ponderosa/data/api/HomeSeerService.java
@@ -17,7 +17,7 @@ public interface HomeSeerService {
   Observable<Result<HSDeviceControlResponse>> deviceControls(@Query("ref") List<String> references);
 
   @GET("JSON?request=controldevicebyvalue") //
-  Observable<Result<HSDevicesResponse>> controlDevice(@Query("ref") String reference, @Query("value") Integer value);
+  Observable<Result<HSDevicesResponse>> controlDevice(@Query("ref") String reference, @Query("value") Long value);
 
   @GET("JSON?request=getevents") //
   Observable<Result<HSEventsResponse>> events();
